requirements.txt
opencv-python==4.6.0.66
Pillow==9.3.0
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0